Line 53
Sorry I realized I just proposed a change on an old branch, so apologies if you read this twice. Like I said on that one, lines 52 and 53 are the same. Which results in overwriting the desired candidate to their rank. So rather than preferences[2][2] = preferences[3][2] = preferences[4][2] = 0, it = 2. Since there is no value within  preferences[2][3] = preferences[3][3] = preferences[4][3], its output is "Invalid vote." rather than the desired "3 3 1 0 " that the _init_.py file suggests.